Updated outputSpeciesJsonForHadoop to run as an external python module
If you run it as a GP tool from within ArcGIS (either in process or out
of process) then there are a lot of memory leaks and it eventually
causes ArcGIS to crash. However, if you run the same script as a Python
module from the command line, then there are no memory leaks.